Invoice, Order, Item Removal
The purchase activity is currently connected to the browse recycler. You can complete the form and it will generate an order entry for the seller and an invoice entry for the user. Invoice activity and order activity in progress. Fragment will likely be a simple button back to main menu if possible.